LITMUS: don't let rt_domain deadlock in global plugins
On the T2000, GSN-EDF would eventually deadlock. This was caused by a deadlock
involving the release and ready locks of its rt_domain.

Usually, the dependency in GSN-EDF is:

	function		held  -> required
	=========================================
	enqueue: 		ready -> release
	gsnedf_job_release:	      -> ready
	arm_release_timer:	      -> release

So far so good. The problem arose when hrtimer detected that the timer being
armed had already expired (must have been just a few nanoseconds) and decided
to call the callback directly WHILE HOLDING THE RELEASE LOCK. Which led to the
following result:

	function		held    -> required	CPU
	====================================================
	enqueue: 		ready   -> release	5
	gsnedf_job_release:	release -> ready	6

We avoid this problem now by dropping the release lock in arm_release_timer()
before calling the job release callback.