Used miscdevice API for sched_trace
This patch changes sched_trace.c to use the miscdevice API
instead of doing all the cdev management ourselves. This remove a
chunk of code and we get sysfs / udev integration for free.

On systems with default udev rules, this will result in a /dev/litmus/log
device being created automatically.