Bugfix: PSN-EDF should only requeue tasks that are not scheduled
Requeue a task that is already scheduled will cause it to be effectively
in the runqueue twice since scheduled tasks are conceptually the head
of the queue. If a task is still scheduled, then schedule() will do the
right thing and do the requeuing if necessary.

This fixes crashes reported by Glenn and Andrea.