Make TRACE() buffer size configurable
Let the user choose an appropriate buffer size (instead of scaling
with NR_CPUS).  The kfifo api requires the buffer to be a power of
two, so enforce this constraint in the configuration.

This fixes a previously-existing compile-time error for values of
NR_CPU that are not a power of two.

Based on a patch by Mac Mollison <mollison@cs.unc.edu>.