log job number in TRACE_TASK() and TRACE_CUR()
For some problems it can be helpful to know which job of a task
generated a log message. This patch changes TRACE_TASK to add :<jobno>
to the existing (<comm>/<pid>) tag.

The result is a trace such as the following, in which the third job of
rtspin/1511 completes and the fourth job is added to the release
queue.

  137615 P0: (rtspin/1511:3) job_completion().
  137616 P0: (rtspin/1511:4) add_release(), rel=262013223089
  137617 P0: (rtspin/1511:4) scheduled_on = NO_CPU

The job number for non-real-time tasks is always zero.