Feather-Trace: register devices with sysfs
This patch implements support for Feather-Trace devices to use the sysfs
file system and, consequently, udev support.

This allows us to allocate major/minor numbers for Feather-Trace
devices dynamically, which is desirable because our old static
allocations tend to create conflicts on modern distributions and/or
when there are many cores.