Improve FMLP queue management.
The next owner of a FMLP-protected resource is dequeued from
the FMLP FIFO queue by unlock() (when the resource is freed by
the previous owner) instead of performing the dequeue by the next
owner immediately after it has been woken up.

This simplifies the code a little bit and also reduces potential
spinlock contention.