GSN-EDF: do not requeue jobs without budget
This patch changes how preemptions of jobs without
budget work. Instead of requeuing them, they are now
only added if they are not subject to budget enforcement
or if they have non-zero budget. This allows us to process
job completions that race with preemptions.

This appears to fix a BUG in budget.c:65 reported by Giovani Gracioli.