Re-implement synchronous task set release support
The old implementation had the problem that the first job release of all
tasks was handled on the *same* CPU (that triggered the synchronous task set
release). This could result in significant latency spikes.

The new implementation uses a completion per task. Futher, each task programs
its own release, which ensures that it will be on the proper processor.