litmus: get rid of unused RT_F_EXIT_SEM flag
This patch removes the RT_F_EXIT_SEM flag. All code paths
depending on it being true are assumed to be unreachable
and removed.

The 'flags' field in struct rt_params is left as-is for
use by specific schedulers. For example, sched_pfair
defines a custom flag RT_F_REQUEUE within the 'flags'
field.

Signed-off-by: Manohar Vanga <mvanga@mpi-sws.org>