PFAIR: do not double-delete from heap on sync. release
The PFAIR code was no longer working correctly with the synchronous
release code. This caused a crash in bheap_delete(). With this patch,
the drop_all_references() is avoided until complete_job() calls
into the scheduler(), which avoids the double deletion.

This still looks quite brittle to me, but seems to fix the bug (or at
least its obvious symptoms) in my test setup.