Make LITMUS^RT the highest-priority scheduling class again
The existing LITMUS^RT code assumes that it is called as the
highest-priority scheduling class (this was true before stock Linux
gained stop_sched_class). The port to 3.10 changed this, which can
result in rare crashes. This patch makes LITMUS^RT the top scheduling
class again.

Having higher priority than stop_sched_class makes no real sense, but
this was the case in 3.0 also and is required for the existing code to
work correctly.