litmus/ctrldev: Fix stale readouts on ARMv7
For ARMv7 __S011 only declares the access bits, not the caching
policy. As a result the userspace mapping would bypass the cache,
causing the kernel to occasionally read stale entries from the cache.
Switch to using PAGE_SHARED instead, which assumes correct caching
defaults.

[With kind assistance by Alexander Zuepke <alexander.zuepke@hs-rm.de>,
who pointed out the correct direction to look for the caching bug.]

Signed-off-by: Roy Spliet <rspliet@mpi-sws.org>