P-FP: remove agent priorities from regular PCP
Due to prior copy&paste reuse, the PCP used agent priorities like in
the DPCP. When using *only* the PCP by itself this is not a problem
(all tasks are similarly treated, which is equivalent to none of the
tasks using agent priorities), but it does look strange and also
prevents the PCP from being used properly together with the DPCP, as
it makes all local resources have DPCP-like ceilings.

This patch removes agent priorities from the PCP. Since the effective
priority of a task is given as a parameter to the PCP implementation,
this does not affect the DPCP implementation.