P-FP: document that PCP and DPCP do not handle remote unlock
It is currently not possible to unlock a PCP or DPCP from the "wrong"
core.  In normal operation, this never happens. In a rare corner case,
however, it is conceivable to occur if the following sequence of
events occur:

1) {D}PCP lock()
2) task transitions to background mode
3) task migrates to different core as a best-effort task
4) {D}PCP unlock() <- remote unlock attempt

Fixing this is a bit tricky. For now, let's at least document the
shortcoming and avoid failing silently by adding a corresponding
assertion.