SRP: document that the SRP does not handle remote unlock
SRP resources are per-processor resources. The code assumes that the
unlocking task is on the "correct" core. This is generally true under
the real-time plugins that support the SRP, but may not be true if a
lock-holding task transitioned out of real-time mode.

This patch documents this shortcoming in the code and adds an
invariant check to avoid silent failures.