Move domain intialization to non-atomic part of plugin switching path
The procfs library functions make allocations with GFP_KERNEL, which
is not permitted if we call them from atomic contexts (i.e., from
within a stop_machine callback).

This patch moves the activation and deactivation of the scheduling
domain procfs integration before and after the call to stop_machine().
The activation is still serialized with regard to concurrent switching
attempts due to the plugin_switch mutex.