P-FP/MPCP&FMLP+: move wake_up_task() outside critical section
The spin lock protecting each mutex is not required to wake up a task,
so drop it before waking the new lock holder.  This prevents the
lock() operation from blocking up to 20,000 cycles (on ARM) waiting
for intra-processor communication, without changing the worst case of
the unlock operation.

Additional style fixes included.